Implemented subprocess system and converted Shake to work as a subprocess.
